support for parse-float / parse to number (#4644)
* support for parse-float / parse to number

* fix example

* better text

* extra code

+++ pxtsim/libgeneric.ts
@@ -336,7 +336,7 @@ namespace pxsim {
         }
 
         export function toNumber(s: string) {
-            return parseInt(s);
+            return parseFloat(s);
         }
 
         // TODO check edge-conditions
